[TASK] Migration path from tt_address_group to sys_category
In tt_address 3.0, the Category API is used instead of the built-in
address groups in tt_address. Therefore, a wizard is added to the
install tool that converts tt_address_group items to sys_category
items (respecting language), writes the MM relations and also
updates the group selection in the FlexForm.

Releases: 3.0
Resolves: #71430